[SP-139] refactor: instantmeeting 패키지명 수정
