+++ src/main/java/com/cupid/jikting/meeting/service/InstantMeetingService.java
@@ -1,6 +1,6 @@
-package com.cupid.jikting.instantmeeting.service;
+package com.cupid.jikting.meeting.service;
 
-import com.cupid.jikting.instantmeeting.dto.InstantMeetingResponse;
+import com.cupid.jikting.meeting.dto.InstantMeetingResponse;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
